requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
httpx==0.27.0
pydantic==2.6.3
python-multipart==0.0.9
